test(compat-angular): rename angular-zorro suite to ng-zorro-antd
