+++ compat/angular/src/ng-zorro-antd/slider.ts
