+++ compat/angular/tests/ng-zorro-antd/date.spec.ts
@@ -1,6 +1,6 @@
 import { setFieldValue } from '@letsrunit/playwright';
 import { expect, test } from '@sand4rt/experimental-ct-angular';
-import { NzDatepickerFixture, NzDateRangePickerFixture } from '../../src/angular-zorro/date';
+import { NzDatepickerFixture, NzDateRangePickerFixture } from '../../src/ng-zorro-antd/date';
 import { registerHtmlDump } from './debug-html';
 
 registerHtmlDump(test);
