+++ compat/angular/tests/ng-zorro-antd/slider.spec.ts
@@ -1,6 +1,6 @@
 import { setFieldValue } from '@letsrunit/playwright';
 import { expect, test } from '@sand4rt/experimental-ct-angular';
-import { NzSliderFixture } from '../../src/angular-zorro/slider';
+import { NzSliderFixture } from '../../src/ng-zorro-antd/slider';
 import { registerHtmlDump } from './debug-html';
 
 registerHtmlDump(test);
